requirements.txt
typer>=0.12.0
mem0ai>=0.0.8
pydantic>=2.7.0
networkx>=3.2.0
pytest>=8.0.0
sentence-transformers>=2.2.0

